Implementented basic (and buggy) pointer support.
Pointer suppor is implemented with a color-picking algorithm to map the pointer location to objects. Each Renderable now has a "uid" which is nothing more than an RGB tuple, and an updated paint method to support different render types. When we request a PAINT_UID render type, objects paint themselves with a flat color corresponding to the uid. Subobjects are painted similarly. When we want to know what the mouse is over, we simply request the color of the pixel at its location and then search through the objects.

Color picking depends on an exact color match, which for some reason SDL is not providing. Some Lights are correctly colored, but others are not (and I suspect are just barely incorrect)

Although the code mostly works