Replace Wiimote-as-keyboard with rudimentary Wiimote-as-joystick
Code crashes on the Wii after a few minutes, but works fairly well otherwise. All old keypresses are mapped to their joystick equivalent. Classic controller mapping is also included, but this is currently failing for the d-pad (HAT) which is rotated to the left for the Wiimote, but must not be rotated when using the classic controller.