Initial modification to new render system.
Rather than having a "Renderer" thread that is responsible for periodically sending out a paint() request, the new system has the ScreenRenderer be a Renderable that can get dirty like anything else. When the renderable detects that it is dirty, it repaints itself. When other objects are made dirty, they notify their parent to become dirty as well (propogating dirt up the tree). This maintains an upper bound of renders per second and also provides for no lower bound (if nothing ever gets dirty, the ScreenRenderer never gets dirty and never does a blit).

Additionally, the notion of a "CachedRenderable" has been broken out to its own class for now. Given the problem that NOT caching has currently (we never free all the once-off surfaces we create!), it may be folded back into Renderable.

The code is in a semi-working condition. Although it compiles fine, it does not work with both the Cursor and LightsOutManager code in main.cpp. Unless one of the two are commented out X kills the program as soon as the mouse enters. Clearly there is some kind of interaction between the two classes and the ScreenRenderer.