Fixed alpha blending of cursor surface onto game.
Problem was a result of SDL_SRCALPHA being implicitely set on the PNG cursor texture. RGBA-RGBA blits with this property do not modify the destination alpha channel, meaning that even though the final cursor surface contained the pointer image, it was totally transparent when blitted by ScreenRenderer. Manually disabling this flag causes the alpha from the PNG to replace the alpha in the cursor surface, which gives us appropriate transparency with ScreenRenderer.