Fixed a stupid reason for the Wii crash.
Program was crashing when trying to access resources. Earlier on I had a lot of trouble with fopen, and resolved it with a chdir to the lightsout directory. However, with my test being put into a different directory it couldn't find the files ;) Removing that chdir did the trick. Don't ask me why it was required earlier but is hunky-dory without it now...