Change from pointer-style to NES-style control.
Joystick events are interpreted with UP/DOWN/LEFT/RIGHT being correct when the wiimote is held NES-style, and when using the classic controller. Since we also want to support both devices it is necessary to switch from the pointer-style layout to the NES-style.

This also inadvertantly fixes a problem with Wiimote pointer support. With the A button being mapped to both the left mouse button *and* a joystick button, pressing it would duplicate the select event. Now when used in pointer mode only a single select will occur, and similarly in NES mode.