Change behavior of SDL_QUIT and SDL_USEREVENT
The former now gracefully exits while the latter kills. This was implemented to fix a crash when the window's close button was clicked (which sends SDL_QUIT)

This also implements a timeout for EventPublisher. If all objects are not unregistered after a certian ammount of time, SDL_USEREVENT is now sent to each to kill the thread and the publisher exits itself. This should fix hangs on exit due to threads not ending when they should.